% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baker}
\alias{baker}
\title{Citation data between social work journals for the 1985-86 period}
\format{An object of class \code{matrix} with 20 rows and 20 columns.}
\usage{
data(baker)
}
\description{
This example consists of the citation data between social 
work journals for the 1985-86 period, collected and analyzed 
in Baker (1992) – see Table 3 below. The dataset covers 20 
journals and the valued directional ties represents the total
number of (discrete) directional citations between and within
journals. Tie values range between 2 and 356, with a mean,
median and standard deviation of, respectively, 33.6, 16,
and 48.9. Apart from Baker’s original analysis, the dataset
has been used as an example to demonstrate core-periphery
heuristics (Borgatti & Everett, 2000), 2-mode generalized
blockmodeling (Doreian, Batagelj, & Ferligoj, 2005, p. 267ff),
and the direct and indirect deviational 
blockmodeling approach (Nordlund, 2016),
latter studies that exclude the diagonal values.
}
\examples{
data(baker)
Transforming it to matrix format
baker <- as.matrix(baker)
putting zeros on the diagonal
diag(baker) <- 0 
}
\references{
\enc{Baker, D. R.}{Baker, D. R.} (1992). A Structural Analysis of Social Work Journal Network: 1985-1986. Journal of Social Service Research, 15(3-4), 153-168. doi: 10.1300/J079v15n03_09
}
\keyword{datasets}
