% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderImage.R
\encoding{UTF-8}
\name{reorderImage}
\alias{reorderImage}
\title{Reordering an image matrix of the blockmodel (or an error matrix based on new and old partition}
\usage{
reorderImage(IM, oldClu, newClu)
}
\arguments{
\item{IM}{An image or error matrix.}

\item{oldClu}{Old partition.}

\item{newClu}{New partition, the same as the old one except for class labeles.}
}
\value{
Reorder matrix (rows and columns are reordred).
}
\description{
Reorders an image matrix of the blockmodel (or an error matrix based on new and old partition.
The partitions should be the same, except that classes can have different labels.
It is useful when we want to have a different order of classes in figures and then also in image matrices. Currently it is only suitable for one-mode blockmodels.
}
\references{
\enc{Žiberna, A.}{Ziberna, A.} (2007). Generalized Blockmodeling of Valued Networks. Social Networks, 29(1), 105-126. doi: 10.1016/j.socnet.2006.04.002

\enc{Žiberna, A.}{Ziberna, A.} (2008). Direct and indirect approaches to blockmodeling of valued networks in terms of regular equivalence. Journal of Mathematical Sociology, 32(1), 57-84. doi: 10.1080/00222500701790207
}
\seealso{
\code{\link{critFunC}}, \code{\link{plot.mat}}, \code{\link{clu}}, \code{\link{IM}}, \code{\link{err}}
}
\author{
Ales Ziberna
}
\keyword{manip}
