% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ss}
\alias{ss}
\alias{ad}
\title{Sum of Squared deviations from the mean and sum of Absolute Deviations from the median}
\usage{
ss(x)

ad(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Sum of Squared deviations from the mean or sum of Absolute Deviations from the median.
}
\description{
Functions to compute Sum of Squared deviations from the mean and sum of Absolute Deviations from the median.
}
\author{
\enc{Aleš Žiberna}{Ales Ziberna}
}
\keyword{univar}
