% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crand.R, R/crand2.R, R/rand.R, R/rand2.R
\encoding{UTF-8}
\name{crand}
\alias{crand}
\alias{crand2}
\alias{rand}
\alias{rand2}
\title{Comparing partitions}
\usage{
crand(tab)

crand2(clu1, clu2)

rand(tab)

rand2(clu1, clu2)
}
\arguments{
\item{tab}{A contingency table obtained as a table(clu1, clu2).}

\item{clu1}{The two partitions to be compared, given in the form of vectors, where for each unit a cluster membership is given.}

\item{clu2}{The two partitions to be compared, given in the form of vectors, where for each unit a cluster membership is given.}
}
\value{
The value of Rand Index (corrected/adjusted for chance)
}
\description{
Rand Index and Rand Index corrected/adjusted for chance for comparing partitions (Hubert & Arabie, 1985).
The names of the clusters do not matter.
}
\references{
Hubert, L., & Arabie, P. (1985). Comparing Partitions. Journal of Classification, 2(1), 193-218.
}
\author{
\enc{Aleš Žiberna}{Ales Ziberna}
}
\keyword{cluster}
